Removed all warps
The world map that began in February of 2022 on the Safe Survival minecraft server has been closed. The server will reopen with a new map on Jan 26, 2025. All recorded warp locations have been wiped, and the coordinates of the new Spawn and warps will begin being added anew once the new world map is opened.